requirements.txt
# Core ML
torch==2.6.0+cu124
torchvision==0.21.0+cu124
transformers==5.4.0
sentence-transformers==5.3.0
lightgbm==4.6.0
scikit-learn==1.8.0
numpy==2.4.3
pandas==3.0.1

# Utilities
joblib==1.5.3
psutil==7.2.2
python-dotenv==1.2.2

# Database
psycopg2-binary==2.9.11

# API (FastAPI backend)
fastapi
uvicorn[standard]
pydantic
